// internal/client/src/models/kibana.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: alerts.go

export interface KibanaWatcherLogResult {
  execution_time: string;
}
export interface KibanaWatcherLogSource {
  result: KibanaWatcherLogResult;
  watch_id: string;
}
export interface KibanaWatcherLog {
  _id: string;
  _source: KibanaWatcherLogSource;
  sort: any[];
  coordinates: KibanaLogCoordinates;
}
export type KibanaWatcherLogs = (KibanaWatcherLog | undefined)[];

//////////
// source: errors.go

export interface KibanaErrorLogSource {
  correlationId: string;
  tcr: string;
  environment: string;
  httpStatus: number /* int32 */;
  message: string;
  microservice: string;
  errorMessage: string;
  '@timestamp': string;
}
export interface KibanaErrorLog {
  _id: string;
  _source: KibanaErrorLogSource;
  sort: any[];
  coordinates: KibanaLogCoordinates;
}
export interface KibanaLogErrorComparable {
  KibanaErrorLog?: KibanaErrorLog;
}
export type KibanaErrorLogs = (KibanaErrorLog | undefined)[];

//////////
// source: kibana.go

export interface KibanaLogCoordinates {
  error: any /* similarity.Coordinate */;
}
export interface KibanaLog {
  _id: string;
  _source: { [key: string]: any};
  sort: any[];
  coordinates: KibanaLogCoordinates;
}
export type KibanaLogs = (KibanaLog | undefined)[];
export interface KibanaHits {
  hits: KibanaLogs;
  total: number /* float64 */;
}
export interface KibanaSearchResult {
  hits: KibanaHits;
}
export interface KibanaClient {
  URL: string;
  Username: string;
  Password: string;
}